+++ src/query/storages/fuse/src/operations/common/generators/mutation_generator.rs
@@ -1,0 +1,111 @@
+// Copyright 2021 Datafuse Labs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::sync::Arc;
+
+use databend_common_exception::ErrorCode;
+use databend_common_exception::Result;
+use databend_common_expression::TableSchema;
+use databend_common_metrics::storage::*;
+use databend_storages_common_table_meta::meta::ClusterKey;
+use databend_storages_common_table_meta::meta::TableSnapshot;
+use log::info;
+use uuid::Uuid;
+
+use crate::operations::common::ConflictResolveContext;
+use crate::operations::common::SnapshotGenerator;
+use crate::statistics::merge_statistics;
+use crate::statistics::reducers::deduct_statistics_mut;
+
+#[derive(Clone)]
+pub struct MutationGenerator {
+    base_snapshot: Arc<TableSnapshot>,
+    conflict_resolve_ctx: ConflictResolveContext,
+}
+
+impl MutationGenerator {
+    pub fn new(base_snapshot: Arc<TableSnapshot>) -> Self {
+        MutationGenerator {
+            base_snapshot,
+            conflict_resolve_ctx: ConflictResolveContext::None,
+        }
+    }
+}
+
+impl SnapshotGenerator for MutationGenerator {
+    fn generate_new_snapshot(
+        &self,
+        schema: TableSchema,
+        cluster_key_meta: Option<ClusterKey>,
+        previous: Option<Arc<TableSnapshot>>,
+    ) -> Result<TableSnapshot> {
+        let default_cluster_key_id = cluster_key_meta.clone().map(|v| v.0);
+
+        let previous =
+            previous.unwrap_or_else(|| Arc::new(TableSnapshot::new_empty_snapshot(schema.clone())));
+        match &self.conflict_resolve_ctx {
+            ConflictResolveContext::AppendOnly(_) | ConflictResolveContext::None => {
+                return Err(ErrorCode::Internal(
+                    "conflict_resolve_ctx should not be AppendOnly in MutationGenerator",
+                ));
+            }
+            ConflictResolveContext::ModifiedSegmentExistsInLatest(ctx) => {
+                if let Some((removed, replaced)) =
+                    ConflictResolveContext::is_modified_segments_exists_in_latest(
+                        &self.base_snapshot,
+                        &previous,
+                        &ctx.replaced_segments,
+                        &ctx.removed_segment_indexes,
+                    )
+                {
+                    info!("resolvable conflicts detected");
+                    metrics_inc_commit_mutation_modified_segment_exists_in_latest();
+                    let new_segments = ConflictResolveContext::merge_segments(
+                        previous.segments.clone(),
+                        ctx.appended_segments.clone(),
+                        replaced,
+                        removed,
+                    );
+                    let mut new_summary = merge_statistics(
+                        &ctx.merged_statistics,
+                        &previous.summary,
+                        default_cluster_key_id,
+                    );
+                    deduct_statistics_mut(&mut new_summary, &ctx.removed_statistics);
+                    let new_snapshot = TableSnapshot::new(
+                        Uuid::new_v4(),
+                        &previous.timestamp,
+                        Some((previous.snapshot_id, previous.format_version)),
+                        schema,
+                        new_summary,
+                        new_segments,
+                        cluster_key_meta,
+                        previous.table_statistics_location.clone(),
+                        previous.index_info_locations.clone(),
+                    );
+                    return Ok(new_snapshot);
+                }
+            }
+        }
+        metrics_inc_commit_mutation_unresolvable_conflict();
+        Err(ErrorCode::UnresolvableConflict(format!(
+            "conflict resolve context:{:?}",
+            self.conflict_resolve_ctx
+        )))
+    }
+
+    fn set_conflict_resolve_context(&mut self, ctx: ConflictResolveContext) {
+        self.conflict_resolve_ctx = ctx;
+    }
+}
