+++ src/query/storages/fuse/src/operations/common/generators/snapshot_generator.rs
@@ -1,0 +1,42 @@
+// Copyright 2021 Datafuse Labs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::sync::Arc;
+
+use databend_common_exception::Result;
+use databend_common_expression::TableSchema;
+use databend_storages_common_table_meta::meta::ClusterKey;
+use databend_storages_common_table_meta::meta::TableSnapshot;
+
+use crate::operations::common::ConflictResolveContext;
+
+#[async_trait::async_trait]
+pub trait SnapshotGenerator {
+    fn set_conflict_resolve_context(&mut self, ctx: ConflictResolveContext);
+
+    async fn fill_default_values(
+        &mut self,
+        _schema: TableSchema,
+        _snapshot: &Option<Arc<TableSnapshot>>,
+    ) -> Result<()> {
+        Ok(())
+    }
+
+    fn generate_new_snapshot(
+        &self,
+        schema: TableSchema,
+        cluster_key_meta: Option<ClusterKey>,
+        previous: Option<Arc<TableSnapshot>>,
+    ) -> Result<TableSnapshot>;
+}
